# Complete Child Issue #7: Performance Optimization & Documentation