--- a/test-build/themes/compose/.gitignore
+++ b/test-build/themes/compose/.gitignore
@@ -1,7 +1,0 @@
-public/
-node_modules/
-.DS_Store
-package-lock.json
-yarn.lock
-__generated__
-.hugo_build.lock