--- a/test-build/themes/compose/assets/js/custom.js
+++ b/test-build/themes/compose/assets/js/custom.js
@@ -1,1 +1,0 @@
-// add custom js in this file