--- a/test-build/themes/compose/assets/sass/_custom.sass
+++ b/test-build/themes/compose/assets/sass/_custom.sass
@@ -1,4 +1,0 @@
-// add customs styles and general overrides here
-// due to the cascading nature of css, if you try to override theme css variables in this file, those changes will not apply. Instead, override css variables in the `override.sass` file
-// we recommend not editing this file directly. Instead, create an `assets/sass/_custom.sass` file at the root level of your site.
-// if you edit this file directly, you will have to resolve git conflicts when and if you decide to pull changes we make on the theme