--- a/test-build/themes/compose/exampleSite/static/admin/.gitignore
+++ b/test-build/themes/compose/exampleSite/static/admin/.gitignore
@@ -1,2 +1,0 @@
-index.html
-assets/